Only clean settings for On-prem

+++ internal/database/orgs.go
@@ -9,6 +9,7 @@ import (
 	"github.com/jackc/pgconn"
 	"github.com/keegancsmith/sqlf"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -331,8 +332,11 @@ func (o *orgStore) HardDelete(ctx context.Context, id int32) (err error) {
 		return err
 	}
 
-	if _, err := tx.Handle().DB().ExecContext(ctx, "DELETE FROM settings WHERE org_id=$1", id); err != nil {
-		return err
+	// For On-Prem, notebooks are enabled through settings, so this entry needs to be removed before an org is deleted.
+	if !envvar.SourcegraphDotComMode() {
+		if _, err := tx.Handle().DB().ExecContext(ctx, "DELETE FROM settings WHERE org_id=$1", id); err != nil {
+			return err
+		}
 	}
 
 	res, err := tx.Handle().DB().ExecContext(ctx, "DELETE FROM orgs WHERE id=$1", id)
